pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "FastAPI template with DI, JWT auth, unit and e2e tests"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "pydantic",
    "PyJWT",
    "python-dotenv",
    "uvicorn[standard]",
]

[dependency-groups]
dev = [
    "httpx",
    "mypy",
    "pre-commit",
    "pyright",
    "pytest",
    "pytest-cov",
    "ruff",
]

[tool.uv]
default-groups = ["dev"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
explicit_package_bases = true
ignore_missing_imports = true
